+++ packages/@expo/cli/CHANGELOG.md
@@ -19,7 +19,9 @@
 
 ### 🐛 Bug fixes
 
-- Use specifier-based stable IDs for RSC client boundaries to fix pnpm, package.json exports, and Metro-specific resolution issues ([#41823](https://github.com/expo/expo/pull/41823) by [@kimchi-developer](https://github.com/kimchi-developer))
+- Fix RSC asset files not being included in SSR manifest during production export. ([#41823](https://github.com/expo/expo/pull/41823) by [@kimchi-developer](https://github.com/kimchi-developer))
+- Fix RSC server action HMR in development mode. ([#41823](https://github.com/expo/expo/pull/41823) by [@kimchi-developer](https://github.com/kimchi-developer))
+- Use specifier-based stable IDs for RSC client boundaries to fix pnpm, package.json exports, and Metro-specific resolution issues. ([#41823](https://github.com/expo/expo/pull/41823) by [@kimchi-developer](https://github.com/kimchi-developer))
 - clean up hanging processes at the end of `expo export` better. ([#41692](https://github.com/expo/expo/pull/41692) by [@EvanBacon](https://github.com/EvanBacon))
 - refactor launching Expo Go on Android ([#40020](https://github.com/expo/expo/pull/40020) by [@vonovak](https://github.com/vonovak))
 - only skip dependency validation for `EXPO_NO_DEPENDENCY_VALIDATION=1` ([#40043](https://github.com/expo/expo/pull/40043) by [@kitten](https://github.com/kitten))